Use PropertySelector in plot controls and add grouping to it
