+++ webapp/src/js/components/containers/PlotWithActions.js
@@ -1,15 +1,11 @@
 import React from 'react';
-import Immutable from 'immutable';
-import ImmutablePropTypes from 'react-immutable-proptypes';
 import titleCase from 'title-case';
 import scrollbarSize from 'scrollbar-size';
 import Sidebar from 'react-sidebar';
 
 // Lodash
 import _map from 'lodash/map';
-import _each from 'lodash/map';
 import _reduce from 'lodash/reduce';
-import _filter from 'lodash/filter';
 
 // Mixins
 import PureRenderMixin from 'mixins/PureRenderMixin';
@@ -17,7 +13,6 @@ import ConfigMixin from 'mixins/ConfigMixin';
 import FluxMixin from 'mixins/FluxMixin';
 
 // Material UI
-import Divider from 'material-ui/Divider';
 import {FlatButton} from 'material-ui';
 import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
@@ -25,6 +20,7 @@ import MenuItem from 'material-ui/MenuItem';
 // Panoptes UI
 import SidebarHeader from 'ui/SidebarHeader';
 import Icon from 'ui/Icon';
+import PropertySelector from 'panoptes/PropertySelector'
 
 // Panoptes
 import SQL from 'panoptes/SQL';
@@ -90,23 +86,6 @@ let PlotWithActions = React.createClass({
       label: table.capNamePlural
     }));
 
-    let propertyMenu = [];
-    let i = 0;
-    if (table) {
-      const propertyGroups = this.config.tablesById[table].propertyGroups;
-      _each(propertyGroups, (group) => {
-        if (propertyMenu.length) {
-          propertyMenu.push(<Divider key={i++}/>);
-        }
-        let {id, name} = group;
-        propertyMenu.push(<MenuItem disabled value={id} key={id} primaryText={name}/>);
-        _each(group.properties, (property) => {
-          let {id, name} = property;
-          propertyMenu.push(<MenuItem value={id} key={id} primaryText={name}/>);
-        });
-      });
-    }
-
     let filterButtonLabel = 'Change Filter';
     let decodedQuery = SQL.WhereClause.decode(query);
     let clearFilterButton = null;
@@ -150,13 +129,11 @@ let PlotWithActions = React.createClass({
           </SelectField>
           {table && plotType ?
             _map(plotTypes[plotType].dimensions, (dimension) =>
-              <SelectField value={this.config.tablesById[table].propertiesById[this.props[dimension]] ? this.props[dimension] : null}
-                           key={dimension}
-                           autoWidth={true}
-                           floatingLabelText={titleCase(dimension)}
-                           onChange={(e, i, v) => componentUpdate({[dimension]: v})}>
-                {propertyMenu}
-              </SelectField>
+              <PropertySelector table={table}
+                                key={dimension}
+                                value={this.config.tablesById[table].propertiesById[this.props[dimension]] ? this.props[dimension] : null}
+                                label={titleCase(dimension)}
+                                onSelect={(v) => componentUpdate({[dimension]: v})}/>
             )
             : null }
         </div>
