+++ webapp/src/styles/genomebrowser.scss
@@ -195,6 +195,7 @@
           justify-content: flex-start;
           .label {
             padding-left: 10px;
+            padding-right: 5px;
           }
           .dropdown {
             top: -4px;
